% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixation.R
\name{allele.fixation}
\alias{allele.fixation}
\title{Allele fixation}
\usage{
allele.fixation(pop)
}
\arguments{
\item{pop}{Current population}
}
\value{
TRUE - the population for the final loci has converged or is empty, FALSE otherwise.
}
\description{
Has the population lost all allelic diversity?
Note that the function uses the last 2 allele values for each individual.
Normally fixation would be run for a single loci, so this assumption is
not too limiting.
}
