% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding.R
\name{baseline.breeding}
\alias{baseline.breeding}
\title{Create population available for breeding}
\usage{
baseline.breeding(p, b.table)
}
\arguments{
\item{p}{Current population of individuals as .ppp}

\item{b.table}{The breeding table defined by \code{breeding.table}}
}
\value{
A new population containing only those individuals who are available for breeding.
}
\description{
Given a population and a breeding table, determine which individuals
can breed based on their age class and the probability of breeding.
}
