% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{baseline.survival}
\alias{baseline.survival}
\title{Baseline survival for individuals}
\usage{
baseline.survival(p, s.table, habitat = NA)
}
\arguments{
\item{p}{The current population}

\item{s.table}{The survival probability table}

\item{habitat}{The habitat surface representing the probability of survival at any location.}
}
\value{
The surviving population as a .ppp
}
\description{
Given a population \code{p}, determine the survival to the next generation based on
the baseline probabilities of survival, \code{s.table}, and the habitat.
}
