% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding.R
\name{breed}
\alias{breed}
\title{Breed valid individuals}
\usage{
breed(pts, b.pairs, b.table)
}
\arguments{
\item{pts}{The points representing individuals that are available for
breeding}

\item{b.pairs}{The breeding pairs as a matrix returned from \code{breeding.pairs}.
Each row is <female> <male> index into \code{pts}.}

\item{b.table}{Breeding table.  This is used to extract the number of children to produce and
displacement of children drawn from normal distribution around the mid-point in space
between the male and female parents.}
}
\value{
The children as a .ppp with id,sex,age, allele values updated.
}
\description{
For each closest pair of male/females, create \code{fecund}
 number of children, and place them
    around the midpoint of the location between the parents, displaced
    from this point by \code{displace} as an isotropic placement.
    Once parents are seperated by > \code{max.dist} breeding stops.
}
