% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding.R
\name{breeding.pairs}
\alias{breeding.pairs}
\title{Construct breeding pairs of males/females}
\usage{
breeding.pairs(p, max.dist)
}
\arguments{
\item{p}{The population of valid breeding individuals}

\item{max.dist}{Maximum distance allowed between male/female pairs for breeding}
}
\value{
A 2 column table (female, male) of individual indices of breeding pairs.
}
\description{
Given a set of individuals who are available for breeding, pair
up the males/females.  This is done by finding the first female and matching to closest (geographic)
male.  This pair is then removed from the available pool for breeding, and the process
repeated.  Valid breeding pairs must be no further apart than \code{max.dist}.
}
