% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding.R
\name{breeding.table}
\alias{breeding.table}
\title{Define breeding parameters}
\usage{
breeding.table(m.age.class = c(5, 40, Inf), m.breed = c(0, 0.8, 0.1),
  f.age.class = c(6, 40, Inf), f.breed = c(0.8, 0.9, 0.1), fecund = c(5,
  2), displace = c(3, 0.5))
}
\arguments{
\item{m.age.class}{A vector that defines the values for each male age threshold. These are interpreted
as less than or equal values. The last value should
be Inf.  An interval (lower,upper) is interpreted as lower <= V <= upper.  The
first value is assumed to be the upper value for a range (0,upper).}

\item{m.breed}{A vector of probabilities of breeding, one for each male age class}

\item{f.age.class}{A vector that defines each female age class.}

\item{f.breed}{A vector of probabilities of breeding, one for each female age class}

\item{fecund}{Number of children created if breeding occurs. Defined as two value vector
N(m,sd) drawn from a normal distribution.  Values always >= 0.}

\item{displace}{Displacement of children drawn from normal distribution around the mid-point in space
between the male and female parents.}
}
\value{
Breeding table to be used in \code{baseline.breeding}.  Note some basic checking of age class
matching to number of probabilities is done, but no checking of values.
}
\description{
Breeding occurs between males and females who are spatially located
near each other.  The probability of breeding depends on the age (which may be different
for males and females).  Given a breeding event, fecundity is drawn from a defined normal
distribution.  Children are randomly placed around the mid-point between parents.
}
