% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.tree.R
\name{build.population.graph}
\alias{build.population.graph}
\title{Build a graph for a list of populations}
\usage{
build.population.graph(p.list)
}
\arguments{
\item{p.list}{Population list}
}
\value{
A graph as an igraph object
}
\description{
Create a graph using the results of a model run.
Assumes that p.list is a list of populations as .ppp, probably
created using \code{multiple.step}.
}
