% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{circle.habitat}
\alias{circle.habitat}
\title{Define a circular habitat}
\usage{
circle.habitat(x, y, b)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{b}{Decay rate for any habitat location defined by the distance from origin (0,0)}
}
\value{
Habitat value for each location.  This function is used in conjunction with
\code{\link[spatstat]{as.im}} to produce an image that defines the habitat.
}
\description{
Define a circular habitat using an exponential decay function.
The origin(0,0) has value 1, with habitat locations away from the origin
decaying in value towards zero.
}
\examples{
# Example circle habitat for a rectangular window.
# hab.win is the habitat window
hab.win <- owin(c(-50,50),c(-50,50),unitname="metres")
Z <- as.im(circle.habitat,hab.win,b=0.0005)
plot(Z)

}
