% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{circle.only}
\alias{circle.only}
\title{Define a circular habitat with constant value}
\usage{
circle.only(x, y, dist.inside = 50, inside = 1, outside = 0)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{dist.inside}{Distance from the origin (0,0) with habitat value inside}

\item{inside}{Value of habitat within dist.inside from the origin}

\item{outside}{Value of habitat further than dist.inside from the origin}
}
\value{
Habitat value for each location.  This function is used in conjunction with
\code{\link[spatstat]{as.im}} to produce an image that defines the habitat.
}
\description{
Create a circular habitat with a value of \code{inside} within a defined distance
from the origin(0,0), and \code{outside} elsewhere.
}
\examples{
hab.win <- owin(c(-50,50),c(-50,50),unitname="metres")
Z <- as.im(circle.only,hab.win,dist.inside=10,inside=0.8,outside=0.2)
plot(Z)
}
