% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{circle.pts.density}
\alias{circle.pts.density}
\title{Define a set of circular point habitats}
\usage{
circle.pts.density(x, y, hab.win, sigma = 1, min.d = 0.7)
}
\arguments{
\item{x}{x coordinates defining the point locations}

\item{y}{y coordinates of point locations}

\item{hab.win}{The habitat window as an owin object}

\item{sigma}{Density function argument.  Larger values produce greater spread}

\item{min.d}{Minimum value of final habitat map.}
}
\value{
Habitat value for each location as an im object (\code{\link[spatstat]{as.im}}).
This can be used as the habitat window in spatibm.
}
\description{
Define a habitat locations within a window using a density
function.  The maximum value is 1, and the minimum value in the habitat is
set by the argument min.d.
}
\examples{
# Example habitat with 3 good habitat locations.
# hab.win is the habitat window.  Minimum value is 0.7
hab.win <- owin(c(-50,50),c(-50,50),unitname="metres")
Use 3 point locations within the -50x50 window
x = c(-20,0,20)
y = c(-20,0,20)
# Note that Z is type im.
Z <- circle.pts.density(x,y,hab.win,sigma=5,min.d=0.7)
plot(Z)

}
