% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{convex.home.range}
\alias{convex.home.range}
\title{Estimate home range for a movement parameter}
\usage{
convex.home.range(pop, move.table, timesteps = 5)
}
\arguments{
\item{pop}{The population}

\item{move.table}{The movement table}

\item{timesteps}{The number of times movement is applied to each individual.}
}
\value{
Vector of convex-hull areas of length \code{pop$n} representing the movement for each individual.
}
\description{
Given an initial population and movement table, simulate the movement
of the population, starting at the origin, using \code{move.table} for \code{timesteps}.
The resulting table gives the convex hull area for each individual in the population.  The purpose
of this function is to help relate a estimated organism home range to the movement paramters
for the model.  Note that the window for \code{pop} determines the constraints of movement.
Individuals that move outside the population window are removed.  The model also commences
with the individuals at (0,0), so windows should be centred around the origin.
}
