% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{crowding.survival}
\alias{crowding.survival}
\title{Determine survival due to crowding}
\usage{
crowding.survival(p, c.table, sigma = 1)
}
\arguments{
\item{p}{The current surviving population}

\item{c.table}{The crowding (density) table, as defined by \code{crowding.table}.}

\item{sigma}{The kernel density bandwidth}
}
\value{
The surviving population as a .ppp
}
\description{
For each individual, determine the density of the remaining population
around this individual and then determine the probability of survival as defined in \code{c.table}.
Use this to calculate survival. Density is determined by considering the density around an individual, but
not including this individual in the density calculation.  The bandwidth of the kernel used for the density
calculation is given by \code{sigma}.
}
