% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{crowding.table}
\alias{crowding.table}
\title{Define crowding table}
\usage{
crowding.table(m.age.class = c(5, Inf), m.den.class = c(0.5, 1, 4, Inf),
  m.prob = c(1, 0.9, 0.5, 0.01, 1, 0.9, 0.5, 0.01), f.age.class = c(4, Inf),
  f.den.class = c(0.5, 1, 2, 4, Inf), f.prob = c(1, 0.9, 0.8, 0.1, 0.01, 1,
  0.9, 0.8, 0.1, 0.01))
}
\arguments{
\item{m.age.class}{Male age classes}

\item{m.den.class}{Male density classes}

\item{m.prob}{Male probabilities of survival given any age and density class}

\item{f.age.class}{Female age classes}

\item{f.prob}{Female probabilities of survival given any age and density class}

\item{fden.class}{Female density classes}
}
\value{
The crowding table.  Note some basic error checking is performed.  The resulting table should be visualised
using \code{plot.ct} to confirm the crowding table is correctly defined.
}
\description{
Define the crowding table for the population
This involves defining (for both males and females) the age classes and then
for each age class, the density classes, and for each age class/density Class
the probability of survival given this density measure for the specific age class.
To make the definition somewhat simplier we assume that the number of density classes
are the same for each age class.  This isn't a limitation, and makes it somewhat easier to
define the probabilities, since you just have to have the same number per age class.
To make it simplier (and conceptually easier to think about) we split the male and female definitions.
}
\examples{
ct <- crowding.table()  # Use defaults
plot.crowding.table(ct)

}
