% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding.R
\name{do.breeding}
\alias{do.breeding}
\title{Breed a pair of individuals}
\usage{
do.breeding(x, pts, fecund, displace)
}
\arguments{
\item{x}{A row from the breeding pairs table, giving the male and female ids.}

\item{pts}{The .ppp population of breeding individuals.  This is used to access the (x,y) location
and allele values for the parents.}

\item{fecund}{Number of children created if breeding occurs. Defined as two value vector
N(m,sd) drawn from a normal distribution.  Values always >= 0.}

\item{displace}{Displacement of children drawn from normal distribution around the mid-point in space
between the male and female parents.}
}
\value{
The table of children.  See \code{breed} for details.
}
\description{
Breed a pair of individuals, creating a set of offspring.  These are
placed in space based on \code{displace}.  Medellian genetics are used to determine the
new allele values for the offspring.
}
