% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{F.stat}
\alias{F.stat}
\title{Calculate Wright F statistic for a population}
\usage{
F.stat(p)
}
\arguments{
\item{p}{Population}
}
\value{
Numeric vector with 4 values per loci.  Values are p frequency, H observed,
H expected and the F statistic.
}
\description{
Calculate Wright F statistic for a population
}
\examples{
p <- create_ibm_population()
# The default population has two loci, so there are 2 x 4 values.
print(F.stat(p))

}
