% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixation.R
\name{fixation}
\alias{fixation}
\title{Time to fixation for an spatial individual-based model}
\usage{
fixation(max.gens = 100, curr.pop, move.table, survive.table, breed.table,
  habitat.surface = NA, crowd.table = NA, crowding.sigma = 0,
  max.dist = 5, trace.output = FALSE)
}
\arguments{
\item{max.gens}{Maximum number of generations before the simulation halts.}

\item{curr.pop}{The current population as a .ppp}

\item{move.table}{Movement table as defined by \code{movement.table}.}

\item{survive.table}{Survival table as defined by \code{survival.table}.}

\item{breed.table}{Breeding table as defined by \code{breeding.table}}

\item{habitat.surface}{The habitat (probability) surface. See \code{circle.habitat} for an example, although
this surface could be produced in many ways.  The only requirement is that it must cover the entire window
associated with the population.}

\item{crowd.table}{The crowding table as defined by \code{crowding.table}}

\item{crowding.sigma}{Bandwidth used for density calculation for crowding}

\item{max.dist}{Maximum distance between breeding parents}

\item{trace.output}{TRUE - write out generation details to console}
}
\value{
This function will complete when: (a) The alleles have fixed; (b) The maximum number
of generations has been reached; or (c) The population has died out. To allow a simple
method for distinguishing each of these cases, the model returns a 6 element numeric vector: <TRUE/FALSE fixed>,
<generation>, <Final Population Count>, <Allele Sum> <#Males>,<#Females>.
}
\description{
A simple example of using the \code{single.step} function
to determine the time to fixation of a neutral allele for a model.  Assumes
that the \code{curr.pop} has been suitably initialised. Note that all parameters for
the model, other than \code{max.gens} is the same as the model \code{single.step}.
This runs the model once, whereas because of the stochastic nature of the model, you would
need to run this model repeatedly to obtain a good mean and variance of the time to fixation.
}
