% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.tree.R
\name{layout.as.tree}
\alias{layout.as.tree}
\title{Create a tree-style layout for igraph}
\usage{
layout.as.tree(sp, w = 1, h = 1, w.gap = 1, h.gap = 1)
}
\arguments{
\item{sp}{Graph created using \code{build.population.graph}}

\item{w}{Width of allelebox}

\item{h}{Height of allelebox}

\item{w.gap}{Horizontal gap between boxes in graphical units}

\item{h.gap}{Vertical gap between boxes in graphical units}
}
\value{
The igraph layout
}
\description{
Creates a layout design for a graph that assumes we
are showing a family tree structure.  Creates a layout for boxes to be
drawn using the \code{allelebox} object for each individual.
}
