% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_pop.R
\name{make_rpois_num}
\alias{make_rpois_num}
\title{Generate a Poisson stationary set of N points within the observation
window owin.  The intensity of the points is given by N / area(owin)}
\usage{
make_rpois_num(N, owin)
}
\arguments{
\item{N}{Number of points in the population}

\item{owin}{The observation window}
}
\value{
A .ppp point pattern based on the rpoispp (spatstat) function
}
\description{
Generate a Poisson stationary set of N points within the observation
window owin.  The intensity of the points is given by N / area(owin)
}
\examples{
pts <- make_rpois_num(5,
                      owin=owin(xrange=c(-10,10),
                                yrange=c(-10,10),unitname="metre"))
}
