% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{model.summary}
\alias{model.summary}
\title{Create summary table for a model run}
\usage{
model.summary(p.list, sigma = 1, file.csv = NA)
}
\arguments{
\item{p.list}{A list of point populations (.ppp)}

\item{sigma}{Bandwidth used for calculating density for each generation}
}
\value{
A dataframe, one row per generation, showing summary statistics for the population.
}
\description{
Given a list of populations, probably produced from running
a model (see \code{multiple.step}), create a summary table, one row per generation.
}
\examples{
# Create a default population
p <- create_ibm_population()
# Create bigger window for population to expand
p$window <- owin(c(-50,50),c(-50,50),unitname="metres")
# Use default parameters for behaviour.
pm <- multiple.step(steps=5,curr.pop=p,move.table=movement.table(),
survival.table(),breeding.table(),habitat.surface=NA,crowding.table())
print(model.summary(pm))
}
