% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{move}
\alias{move}
\title{Move the individuals in the population using the defined movement table.}
\usage{
move(p, m.table)
}
\arguments{
\item{p}{The current population as a .ppp object}

\item{m.table}{The movement tablee as defined using \code{movement.table}}
}
\value{
A new population with each individual moved according to the \code{movement.table}.
Individuals that have moved outside of the window defined by the population are removed and a
warning given.  This function should not be directly called by the user.
}
\description{
Moving individuals in the population across space assumes that there is
no preferred direction and therefore space is isotropic. Individuals that move outside the
defined window for the .ppp population are removed and a warning given.
}
