% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{movement.table}
\alias{movement.table}
\title{Define movement parameters for individuals}
\usage{
movement.table(m.age.class = c(5, 10, Inf), m.move = list(c(3, 1), c(5, 1),
  c(7, 2)), f.age.class = c(3, Inf), f.move = list(c(2, 1), c(4, 1)))
}
\arguments{
\item{m.age.class}{A vector that defines the values for each male age threshold. These are interpreted
as less than or equal values. The first value assumes an age from zero to the defined value.}

\item{m.move}{A list with a two element vector for each age class that defines the movement
for this age class.  The movement value is drawn from a normal distribution N(m,sd)}

\item{f.age.class}{A vector that defines each female age class.}

\item{f.move}{A list with a two element vector for each age class.}
}
\value{
The movement table. This is passed to the function \code{move}.
}
\description{
Define the movement parameters for each age class of males and females.  Some simple
checking of parameters is done to ensure the correct number of movement parameters are given for the
number of age classes.  Note that movement assumes an isotropic space with no preferred direction.
}
\examples{
# Define a small male movement for age class 0 - 5, and large for >5
# Female movement is the same for all females.
movement.table(m.age.class=c(5,Inf),m.move=list(c(1,0.1),c(5,2)),
               f.age.class=Inf,f.move=list(c(4,1)))
}
