% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{multiple.step.habitat}
\alias{multiple.step.habitat}
\title{Perform multiple steps (generations) for the population, allowing for a stochastic change in habitat.}
\usage{
multiple.step.habitat(steps = 10, curr.pop, move.table, survive.table,
  breed.table, habitat.list, crowd.table, crowding.sigma = 1, max.dist = 5,
  track.ids = TRUE, trace.output = TRUE)
}
\arguments{
\item{steps}{Number of steps (generations) to excute the model}

\item{curr.pop}{The current population as a .ppp}

\item{move.table}{Movement table as defined by \code{movement.table}.}

\item{survive.table}{Survival table as defined by \code{survival.table}.}

\item{breed.table}{Breeding table as defined by \code{breeding.table}}

\item{habitat.list}{The list of habitats (probability) surfaces with associated N(m,sd).}

\item{crowd.table}{The crowding table as defined by \code{crowding.table}}

\item{crowding.sigma}{Bandwidth used for density calculation for crowding}

\item{max.dist}{Maximum distance between breeding parents}

\item{track.ids}{Do you want to update the individual ids each generation?  This is only required if you
need to produce a family tree.}

\item{trace.output}{Show a text-based trace of the model running in the console.}
}
\value{
A list containing the population for each generation.  The initial population \code{curr.pop}
is the first element in the list.
}
\description{
Perform \code{steps} generations for a population based on their defined breeding, survival,
and crowding parameters and the habitat surface.
The \code{habitat.list} defines a list of habitats with a
N(m,sd) for each habitat used to determine the duration this habitat is observed.
A list of these habitat,N(m,sd) structures defines the temporal pattern of change,
and the order of change.  i.e. a list of ( (H1,N1,(H2,N2), ...) defines an ordering of
habitat surface.  The N(m,sd) is used to determine the time period before moving from Hi to H(i+1).
In addition, there is an implied cycling behavior since when the final (H,N(m,sd)) pattern is
used, the next habitat is (H1,N1).
}
