% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{multiple.step}
\alias{multiple.step}
\title{Perform multiple steps (generations) for the population}
\usage{
multiple.step(steps = 10, curr.pop, move.table, survive.table, breed.table,
  habitat.surface = NA, crowd.table, crowding.sigma = 1, max.dist = 5,
  track.ids = FALSE, trace.output = TRUE)
}
\arguments{
\item{steps}{Number of steps (generations) to excute the model}

\item{curr.pop}{The current population as a .ppp}

\item{move.table}{Movement table as defined by \code{movement.table}.}

\item{survive.table}{Survival table as defined by \code{survival.table}.}

\item{breed.table}{Breeding table as defined by \code{breeding.table}}

\item{habitat.surface}{The habitat (probability) surface. See \code{circle.habitat} for an example, although
this surface could be produced in many ways.  The only requirement is that it must cover the entire window
associated with the population.  This surface is constant over all generations.  See \code{multiple.step.habitat}
for control over a changing habitat for multiple steps.}

\item{crowd.table}{The crowding table as defined by \code{crowding.table}}

\item{crowding.sigma}{Bandwidth used for density calculation for crowding}

\item{max.dist}{Maximum distance between breeding parents}

\item{track.ids}{Do you want to update the individual ids each generation?  This is only required if you
need to produce a family tree.}

\item{trace.output}{Show a text-based trace of the model running in the console.}
}
\value{
A list containing the population for each generation.  The initial population \code{curr.pop}
is the first element in the list.
}
\description{
Perform \code{steps} generations for a population based on their defined breeding, survival,
and crowding parameters and the habitat surface.
}
