% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot.all.gens}
\alias{plot.all.gens}
\title{Plot all generations of a model as points}
\usage{
plot.all.gens(p.list, xlim = NA, ylim = NA, cex = 0.7, pch = 16, ...)
}
\arguments{
\item{p.list}{List of populations as .ppp, probably produced by running a model.}

\item{xlim, ylim}{Required explicitly if used, since \code{plot.ppp} uses the .ppp window
as the default window to scale the plot.}

\item{cex}{Size of points used for plotting}

\item{pch}{Point type}

\item{...}{Other graphical parameters}
}
\description{
Plot all generations of a model run, one plotting window
per generation.  Note that the number of plotting windows needs to be
set prior to calling (probably using \code{par(mfrow=c(r,c),...}).
}
\examples{
# Create a default population
p <- create_ibm_population()
# Create bigger window for population to expand
p$window <- owin(c(-50,50),c(-50,50),unitname="metres")
# Use default parameters for behaviour.
pm <- multiple.step(steps=5,curr.pop=p,move.table=movement.table(),
survival.table(),breeding.table(),habitat.surface=NA,crowding.table())
op <- par(mfrow=c(2,3),mar=c(1,1,0.5,0.5))  # setup graphics
plot.all.gens(pm)
par(op) # reset graphics
}
