% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plot.crowding.table}
\alias{plot.crowding.table}
\title{Plot the crowding table}
\usage{
\method{plot}{crowding.table}(c.t, sex = 1, ...)
}
\arguments{
\item{c.t}{Defined crowding table from \code{crowding.table}.}

\item{...}{Graphical parameters}
}
\description{
Graphically show the crowding table as a plot of age versus crowding density.  The probability
of survival based on density is shown by value and size of symbol.
}
\examples{
ct <- crowding.table()  # Use defaults
plot.crowding.table(ct)
}
