% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.density.R
\name{plot.density.example}
\alias{plot.density.example}
\title{Density plot for a given number of points and bandwidth}
\usage{
\method{plot}{density.example}(win, num.pts = 20, sigma = 1)
}
\arguments{
\item{win}{Spatial window}

\item{num.pts}{Number of points randomly placed in the window}

\item{sigma}{Bandwidth}
}
\description{
This function plots a randomly placed set of \code{num.pts} in
the window \code{win}, and colours the surface with the density using the bandwidth
\code{sigma}.  The density at (0,0) is displayed.  This is useful when understanding
and determining the crowding density and the relationship to bandwidth.
}
\examples{
# Create a window centred on (0,0)
win <- owin(xrange=c(-2,2),yrange=c(-2,2),unitname="metres")
plot.density.example(win,num.pts=10,sigma=1)

}
