% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot.pop.count}
\alias{plot.pop.count}
\title{Plot total population}
\usage{
\method{plot}{pop.count}(p.list)
}
\arguments{
\item{p.list}{List of populations created by running a model.}
}
\description{
Create a single plot showing total number of individuals,
and number of males/females in the population for each generation.
}
