% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.tree.R
\name{plot.population.graph}
\alias{plot.population.graph}
\title{Plot the population as a family tree}
\usage{
\method{plot}{population.graph}(g, pop.series, tree.layout = NA,
  vertex.height = 0.02, vertex.width = 0.04, edge.width = rep(0.02,
  length(E(g))), edge.arrow.size = 0.35, edge.curved = TRUE, ...)
}
\arguments{
\item{g}{igraph created using \code{build.population.graph}.}

\item{pop.series}{The list of populations used to build the population graph.}

\item{tree.layout}{The tree layout used for plotting}

\item{vertex.height}{Vertex box height}

\item{vertex.width}{Vertex box width}

\item{edge.width}{Edge width of connecting lines}

\item{edge.arrow.size}{Arrow size of graph}

\item{edge.curved}{Are the drawn edges curved?}

\item{...}{other parameters to plot.igraph}
}
\description{
Plot the population as a family tree.  Note that prior
to using this the alleles shape needs to be added.  See the example below. Edges
are red from females, black from males. This function is in development, and only
works for small populations and quite a lot of massaging with the parameters and layouts.
}
\examples{
# Create a small population of size 10
N <- 10
p <- create_ibm_population(N)
# Create bigger window for population to expand
p$window <- owin(c(-50,50),c(-50,50),unitname="metres")
# Use default parameters for behaviour.
# Note that we require the ids for a tree, so track.ids=TRUE
pop.series <- multiple.step(steps=5,curr.pop=p,move.table=movement.table(),
survival.table(),breeding.table(),habitat.surface=NA,crowding.table(),
track.ids=TRUE)
g <- build.population.graph(pop.series)

# And now add the shape function - depends on the number of
# alleles, so needs to be done after the graph has been built.

add_shape("alleles", clip=shapes("circle")$clip,
       plot=allelebox,parameters=list(vertex.width=1,vertex.height=1,vertex.alleles=length(V(g))))
plot.population.graph(g,pop.series,tree.layout=layout_with_fr(g),asp=1,vertex.height=0.02,vertex.width=0.04)
# Not run - using spatibm package layout
#
# plot.population.graph(g,pop.series,tree.layout=layout_as_tree(g),asp=1,vertex.height=0.02,vertex.width=0.04)
#
}
