% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plot.survival}
\alias{plot.survival}
\title{Create plot showing probability of survival over time}
\usage{
\method{plot}{survival}(p.seq = seq(from = 1, to = 0.5, by = -0.05),
  years = 1:40, ylim = c(0, 1), cex = 1, ...)
}
\arguments{
\item{p.seq}{Vector sequence of probabilities}

\item{ylim}{Default y axis scale}

\item{...}{Other arguments to \code{plot}}

\item{year}{Ordered Vector sequence of years}
}
\description{
A simple plot showing how the probability of survival reduces
over the years.  This can be helpful when determining parameters for survival, crowding
and the habitat surface.
}
\examples{
plot.survival()
plot.survival(xlim=c(0,20),ylim=c(0,0.5))

}
