% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{points.habitat}
\alias{points.habitat}
\title{Define a set of locations with a circular habitat}
\usage{
\method{points}{habitat}(x, y, pts, dist.inside, inside = 1, outside = 0.8)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{pts}{Two column matrix or dataframe of x and y coordinates for a set of points}

\item{dist.inside}{Distance from each point to set value \code{inside}}

\item{inside}{Value to set for locations within \code{dist.inside} of a point}

\item{outside}{Value to set for locations further than \code{dist.inside} from a point.}
}
\value{
Habitat value for each location.  This function is used in conjunction with
\code{\link[spatstat]{as.im}} to produce an image that defines the habitat.
}
\description{
Given a set of points (x,y), create a circular habitat centred on
each point, with value \code{inside} within \code{dist.inside} and value \code{outside} otherwise.
}
\examples{
pts <- matrix(data=c(-60,3,-80,-77,30,30,25,10,-40,-30),ncol=2,byrow=T)
hab.win <- owin(c(-100,100),c(-100,100),unitname="metres")
Z <- as.im(points.habitat,hab.win,pts,dist.inside=20,inside=1.0,outside=0.9)
plot(Z)
}
