% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{pop.summary}
\alias{pop.summary}
\title{Population summary statistics}
\usage{
pop.summary(pop, gen, sigma = 1)
}
\arguments{
\item{pop}{The population as a .ppp}

\item{gen}{The generation of this population.}

\item{sigma}{The bandwidth to use for calculating point density of individuals}
}
\value{
Numeric vector representing a summary of population properties.
See \code{model.summary} to create a labelled table with this summary data.
}
\description{
Given a single population as .ppp, construct a vector
that summarises the properties of the population.
}
