% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixation.R
\name{popsize.survival}
\alias{popsize.survival}
\title{What population size is viable?}
\usage{
popsize.survival(pop.sizes = 10:20, steps = 20, samples = 2,
  spat.layout = "random", obs.win, prob.females = 0.5,
  age.distribution = c(5, 3), allele.prop = c(0.5), habitat.win, move.table,
  survive.table, breed.table, habitat.list = NA, crowd.table, crowding.sigma,
  max.dist, trace.output = FALSE)
}
\arguments{
\item{pop.sizes}{Vector of population sizes to be run}

\item{steps}{Maximum number of steps (generations)}

\item{samples}{Number of times to repeat model for a given population size}

\item{spat.layout}{String to select layout options.  Currently only
"random" is implemented.}

\item{prob.females}{Probability of females in population (0 - 1)}

\item{age.distribution}{The age distribution for the population.  Drawn from
a normal distribution c(mean, sd) Age must be >= 1.
Generated age is an integer. Separate age distribution properties may be
given for females by setting age.distribution to a four value vector, in which case
age distribution is c(male.mean, male.sd, female.mean,female.sd)}

\item{allele.prop}{Proportion of allele Ai per ith loci for the population.
Well mixed = 0.5, converged = 1.0 (or 0.0)
Alleles are stored in the data frame as a vector of size 2 * number loci.
This is done to make the breeding cycle easier when it is modelled.
For example, with 2 alleles would be represented as (A1,B1,A2,B2) where A1 and A2
are the values on the 2 loci first strand, B1 B2 matching loci on second strand.}

\item{habitat.win}{The final habitat window}

\item{move.table}{Movement table as defined by \code{movement.table}.}

\item{survive.table}{Survival table as defined by \code{survival.table}.}

\item{breed.table}{Breeding table as defined by \code{breeding.table}}

\item{habitat.list}{The list of habitats (probability) surfaces with associated N(m,sd).}

\item{crowd.table}{The crowding table as defined by \code{crowding.table}}

\item{crowding.sigma}{Bandwidth used for density calculation for crowding}

\item{max.dist}{Maximum distance between breeding parents}

\item{trace.output}{TRUE - print out each run information, FALSE quiet}

\item{obs.owin}{Observation window that defines the boundaries of the space where
individuals are placed.}
}
\value{
A dataframe with 2 columns: N, finalN representing
the initial population size and final population size
}
\description{
For a given set of population parameters, run the model for a range
of population sizes and maximum steps (for a number of samples).  Return the initial and
final population size, and the number of steps (in case the final population size is zero).
This can be used to examine the probability of survival based on initial population size.
}
