% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival.table}
\alias{survival.table}
\title{Define survival parameters for individuals}
\usage{
survival.table(m.age.class = c(5, 40, Inf), m.survive = c(0.8, 0.8, 0.1),
  f.age.class = c(3, 40, Inf), f.survive = c(0.8, 0.9, 0.1))
}
\arguments{
\item{m.age.class}{A vector that defines each male age class.
An interval (lower,upper) is interpreted as lower <= V <= upper. The first value is assumed
to have a lower bound of zero.}

\item{m.survive}{A vector of probabilities of survival,  one for each age class}

\item{f.age.class}{A vector that defines each female age class.}

\item{f.survive}{A vector of probabilities, one for each female age class.}
}
\value{
The survival table. This is passed to the function \code{baseline.survival}.
}
\description{
Define the survival parameters for each age class of males and females.  Some simple
checking of parameters is done to ensure the correct number of parameters are given for the
number of age classes.  Note that survival is measured as a probability (although possibly modified by the habitat).
}
